Put beta component in seperate package
