+++ .eslintignore
@@ -6,5 +6,4 @@ config/
 coverage/
 lib
 es
-*/asc-ui/beta
 examples
