+++ .gitignore
@@ -8,7 +8,6 @@ node_modules
 /build
 lib
 es
-*/asc-ui/beta
 /storybook-static
 .sb
 packages/**/*.d.ts
