+++ package.json
@@ -6,13 +6,11 @@
   ],
   "scripts": {
     "prebuild": "npm run build --workspace=@amsterdam/asc-assets",
-    "build": "npm run clean:generated && npm run build:ts && npm run build:ts:es && npm run build:ts:beta",
+    "build": "npm run clean:generated && npm run build:ts && npm run build:ts:es",
     "build:watch": "npm run clean:generated && concurrently \"npm run build:ts -- --watch\" \"npm run build:ts:es -- --watch\"",
-    "build:watch:beta": "npm run clean:generated && npm run build:ts:beta -- --watch",
     "build:storybook": "npm run build --workspace=stories",
-    "build:ts": "tsc -b packages/asc-assets packages/asc-ui",
-    "build:ts:es": "tsc -b packages/asc-assets/tsconfig.es.json packages/asc-ui/tsconfig.es.json",
-    "build:ts:beta": "tsc -b packages/asc-ui/tsconfig.beta.json",
+    "build:ts": "tsc -b packages/asc-assets packages/asc-ui packages/asc-ui-beta",
+    "build:ts:es": "tsc -b packages/asc-assets/tsconfig.es.json packages/asc-ui/tsconfig.es.json packages/asc-ui-beta/tsconfig.es.json",
     "build:netlify": "npm run build && npm run build:storybook",
     "clean:generated": "find ./packages -name \"*.d.ts\" -type f -delete && find ./packages -name \"*.tsbuildinfo\" -type f -delete",
     "lint": "concurrently \"npm run lint:css\" \"npm run lint:js\"",
@@ -20,8 +18,7 @@
     "lint:js": "eslint ./ --ext js,mdx,ts,tsx",
     "publish:canary": "npm run build && lerna publish -c --yes",
     "publish:stable": "npm run build && lerna publish",
-    "start": "concurrently \"npm run build:watch\" \"npm run build:ts:beta\" \"npm run storybook\"",
-    "start:beta": "concurrently \"npm run build:ts\" \"npm run build:ts:es\" \"npm run build:watch:beta\" \"npm run storybook\"",
+    "start": "concurrently \"npm run build:watch\" \"npm run storybook\"",
     "storybook": "npm run start --workspace=stories",
     "stylelint": "stylelint \"./packages/**/*Style.ts\"",
     "test": "jest --collect-coverage",
