+++ packages/asc-ui-beta/src/index.ts
@@ -1,4 +1,4 @@
-import { ascBetaTheme } from './theme'
+import ascBetaTheme from './theme/ascBetaTheme'
 import Button from './components/Button/Button'
 
 // eslint-disable-next-line import/prefer-default-export
