+++ packages/asc-ui-beta/tsconfig.es.json
@@ -1,0 +1,8 @@
+{
+  "extends": "./tsconfig.json",
+  "compilerOptions": {
+    "module": "ES2015",
+    "target": "ES2015",
+    "outDir": "./es"
+  }
+}
