--- packages/asc-ui/src/beta/theme/default/colors.ts
@@ -1,43 +1,0 @@
-import type { Theme } from '../../types'
-
-const colors: Theme.ColorTypeInterface = {
-  primary: {
-    main: '#004699',
-    dark: '#00387a',
-  },
-  secondary: {
-    main: '#ec0000',
-    dark: '#bc0000',
-  },
-  tint: {
-    level1: '#ffffff',
-    level2: '#f5f5f5',
-    level3: '#e6e6e6',
-    level4: '#b4b4b4',
-    level5: '#767676',
-    level6: '#323232',
-    level7: '#000000',
-  },
-  support: {
-    valid: '#00a03c',
-    invalid: '#ec0000',
-    focus: '#fec813',
-  },
-  supplement: {
-    purple: '#a00078',
-    pink: '#e50082',
-    orange: '#ff9100',
-    yellow: '#ffe600',
-    lightgreen: '#bed200',
-    darkgreen: '#00a03c',
-    lightblue: '#009dec',
-  },
-  bright: {
-    main: '#ffffff',
-  },
-  error: {
-    main: '#ec0000',
-  },
-}
-
-export default colors
