--- packages/asc-ui/src/beta/theme/default/globalStyle.ts
@@ -1,5 +1,0 @@
-import type { Theme } from '../../types'
-
-const globalStyle: Theme.GlobalStyleType = ''
-
-export default globalStyle
