--- packages/asc-ui/src/beta/theme/default/spacing.ts
@@ -1,5 +1,0 @@
-import type { Theme } from '../../types'
-
-const spacing: Theme.Spacing = 4
-
-export default spacing
