--- packages/asc-ui/src/beta/theme/index.ts
@@ -1,4 +1,0 @@
-import * as ascBetaTheme from './default'
-
-export { default as ThemeProvider } from './ThemeProvider'
-export { ascBetaTheme }
