--- packages/asc-ui/src/beta/types/index.ts
@@ -1,2 +1,0 @@
-// eslint-disable-next-line import/prefer-default-export
-export { Theme } from './Theme'
