--- packages/asc-ui/src/beta/utils/index.ts
@@ -1,12 +1,0 @@
-export {
-  themeColor,
-  srOnlyStyle,
-  breakpoint,
-  svgFill,
-  fluidTypoStyle,
-  typographyStyle,
-  getValueFromTheme,
-  themeSpacing,
-} from './themeUtils'
-export { valueFromObject } from './valueFromObject'
-export { fromProps } from './fromProps'
