+++ stories/package.json
@@ -28,6 +28,7 @@
   "dependencies": {
     "@amsterdam/asc-assets": "^0.35.0",
     "@amsterdam/asc-ui": "^0.35.0",
+    "@amsterdam/asc-ui-beta": "^0.35.0",
     "react": "^17.0.2",
     "styled-components": "^5.3.0"
   }
