+++ stories/src/beta/Button.stories.mdx
@@ -1,7 +1,7 @@
 import { Fragment } from 'react'
 import { Meta, Story, Canvas } from '@storybook/addon-docs'
 import { Icon, ThemeProvider, GlobalStyle } from '@amsterdam/asc-ui'
-import { Button, ascBetaTheme } from '@amsterdam/asc-ui/beta'
+import { Button, ascBetaTheme } from '@amsterdam/asc-ui-beta'
 import styled from 'styled-components'
 import { Download } from '@amsterdam/asc-assets'
